JSON server filled

+++ src/app/modules/mascotas/components/card/card.component.ts
@@ -3,6 +3,7 @@ import { Router } from '@angular/router';
 import { API_PETS } from '@constants/routes/routes';
 import { IMascota } from '@modules/mascotas/interface/mascotas.interface';
 import { MascotasService } from '@modules/mascotas/services/mascotas.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-card',
@@ -23,9 +24,23 @@ export class CardComponent {
   @Output() delete = new EventEmitter();
 
   eliminarMascota(id: number) {
-    console.log(id);
-    this.mascotaService.eliminarMascota(id).subscribe((data) => {
-      this.delete.emit();
+    Swal.fire({
+      title: '¿Está seguro de eliminar la mascota?',
+      text: 'No podrá recuperar la información',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, eliminar',
+      cancelButtonText: 'Cancelar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.mascotaService.eliminarMascota(id).subscribe((data) => {
+          this.delete.emit();
+        });
+      }
     });
+
+    console.log(id);
   }
 }
